Add error handler

+++ controllers/homeController.js
@@ -1,60 +1,60 @@
 const folderQuery = require("../db/folderQuery");
 const fileQuery = require("../db/fileQuery");
 const { supabase } = require("../supabase/supabase");
+const asyncHandler = require("express-async-handler");
 const { v4: uuidv4 } = require("uuid");
 const { validationResult } = require("express-validator");
 const validateFolder = require("../validators/folderValidator");
 const validateFile = require("../validators/fileValidator");
 
-exports.homeGet = async (req, res) => {
+exports.homeGet = asyncHandler(async (req, res) => {
   if (!req.isAuthenticated()) {
     return res.redirect("/");
   }
 
   const currentFolderId = req.params.folderId || null;
 
-  const folders = await folderQuery.getNestedFolders(
-    req.user.id,
-    currentFolderId,
-  );
-  const folderRoute = await folderQuery.getFolderRoute(
-    req.user.id,
-    currentFolderId,
-  );
+  const [folders, folderRoute, parentFolder, files] = await Promise.all([
+    folderQuery.getNestedFolders(req.user.id, currentFolderId),
+    folderQuery.getFolderRoute(req.user.id, currentFolderId),
+    currentFolderId ? folderQuery.getFolder(req.user.id, currentFolderId) : null,
+    fileQuery.getFilesByFolder(req.user.id, currentFolderId),
+  ]);
 
-  const files = await fileQuery.getFilesByFolder(req.user.id, currentFolderId);
+  if (currentFolderId && !parentFolder) {
+    const error = new Error(
+      "The folder you're trying to access does not exist.",
+    );
+    error.status = 404;
+    throw error;
+  }
 
-  const createError = req.session.createError || null;
-  const oldInput = req.session.oldInput || null;
-  const editError = req.session.editError || null;
-  const editTarget = req.session.editTarget || null;
-  const fileError = req.session.fileError || null;
+  const { createError, oldInput, editError, editTarget, fileError } =
+    req.session || {};
 
   // Clean session
-  req.session.createError = null;
-  req.session.oldInput = null;
-  req.session.editError = null;
-  req.session.editTarget = null;
-  req.session.fileError = null;
+  delete req.session.createError;
+  delete req.session.oldInput;
+  delete req.session.editError;
+  delete req.session.editTarget;
+  delete req.session.fileError;
 
   res.render("home", {
-    parentFolder: currentFolderId
-      ? await folderQuery.getFolder(req.user.id, currentFolderId)
-      : null,
-    folders: folders,
-    folderRoute: folderRoute,
-    files: files,
+    parentFolder,
+    folders,
+    folderRoute,
+    files,
     createError,
     oldInput,
     editError,
     editTarget,
     fileError,
   });
-};
+});
 
 exports.createPost = [
   validateFolder,
-  async (req, res) => {
+  asyncHandler(async (req, res) => {
     const errors = validationResult(req);
     const folderName = req.body["folder-name"];
     const parentId = req.params.folderId || null;
@@ -66,31 +66,56 @@ exports.createPost = [
       return res.redirect("/uploads/" + (parentId ? parentId : ""));
     }
 
+    if (parentId) {
+      const folder = await folderQuery.getFolder(req.user.id, parentId)
+      if (!folder) {
+        const error = new Error("Parent folder not found.")
+        error.status = 404;
+        throw error;
+      }
+    }
+
     await folderQuery.createFolder(folderName, req.user.id, parentId);
 
     res.redirect("/uploads/" + (parentId ? parentId : ""));
-  },
+  })
 ];
 
-  exports.deleteFolder = async (req, res) => {
-    const userId = req.user.id;
-    const folderId = req.params.folderId || null;
-    const routes = await folderQuery.getFolderRoute(userId, folderId);
-    const parentId = routes[routes.length - 1].parentId;
+exports.deleteFolder = asyncHandler(async (req, res) => {
+  const userId = req.user.id;
+  const folderId = req.params.folderId || null;
 
-    await folderQuery.deleteFolder(userId, folderId);
-    
-    res.redirect("/uploads/" + (parentId ? parentId : ""));
-  };
+  if (folderId) {
+    const folder = await folderQuery.getFolder(userId, folderId)
+    if (!folder) {
+        const error = new Error("Folder not found.")
+        error.status = 404;
+        throw error;
+    }
+  } 
+
+  const routes = await folderQuery.getFolderRoute(userId, folderId);
+  const parentId = routes[routes.length - 1].parentId;
+
+  await folderQuery.deleteFolder(userId, folderId);
+
+  res.redirect("/uploads/" + (parentId ? parentId : ""));
+})
 
 exports.editNamePost = [
   validateFolder,
-  async (req, res) => {
+  asyncHandler(async (req, res) => {
     const errors = validationResult(req);
     const userId = req.user.id;
     const folderId = req.params.folderId;
+    
+    const folder = await folderQuery.getFolder(userId, folderId)
 
-    if (!folderId) return;
+    if (!folder) {
+        const error = new Error("Folder not found.")
+        error.status = 404;
+        throw error;
+    }
 
     const routes = await folderQuery.getFolderRoute(userId, folderId);
     const parentId = routes[routes.length - 1].parentId;
@@ -111,12 +136,12 @@ exports.editNamePost = [
     await folderQuery.editFolder(userId, folderId, newName);
 
     res.redirect("/uploads/" + (parentId ? parentId : ""));
-  },
+  })
 ];
 
 exports.addFilePost = [
   validateFile,
-  async (req, res) => {
+  asyncHandler(async (req, res) => {
     const errors = validationResult(req);
     const userId = req.user.id;
     const folderId = req.params.folderId || null;
@@ -125,6 +150,15 @@ exports.addFilePost = [
       req.session.fileError = errors.array()[0].msg;
 
       return res.redirect("/uploads/" + (folderId ? folderId : ""));
+    }
+
+    if (folderId) {
+      const folder = await folderQuery.getFolder(req.user.id, folderId)
+      if (!folder) {
+        const error = new Error("Folder not found.")
+        error.status = 404;
+        throw error;
+      }
     }
 
     const { originalname, buffer, mimetype, size } = req.file;
@@ -139,8 +173,7 @@ exports.addFilePost = [
       });
 
     if (error) {
-      console.error("Upload error:", error);
-      return res.status(500).send("Failed to upload");
+      throw new Error("Failed to upload")
     }
 
     const { data: fileURL } = await supabase.storage
@@ -158,18 +191,24 @@ exports.addFilePost = [
     );
 
     res.redirect("/uploads/" + (folderId ? folderId : ""));
-  },
+  })
 ];
 
 exports.editFilePost = [
   validateFolder,
-  async (req, res) => {
+  asyncHandler(async (req, res) => {
     const errors = validationResult(req);
 
     const userId = req.user.id;
     const fileId = req.params.fileId;
 
     const file = await fileQuery.getFile(userId, fileId);
+
+    if (!file) {
+        const error = new Error("File not found.")
+        error.status = 404;
+        throw error;
+    }
 
     const newName = req.body["folder-name"];
 
@@ -187,14 +226,20 @@ exports.editFilePost = [
     await fileQuery.editFile(userId, fileId, newName);
 
     res.redirect("/uploads/" + (file.folderId ? file.folderId : ""));
-  },
+  })
 ];
 
-exports.deleteFile = async (req, res) => {
+exports.deleteFile = asyncHandler(async (req, res) => {
   const userId = req.user.id;
   const fileId = req.params.fileId;
 
   const file = await fileQuery.getFile(userId, fileId);
+
+  if (!file) {
+    const error = new Error("File not found.")
+    error.status = 404;
+    throw error;
+  }
 
   await fileQuery.deleteFile(userId, fileId);
 
@@ -203,13 +248,19 @@ exports.deleteFile = async (req, res) => {
     .remove([file.bucketId]);
 
   res.redirect("/uploads/" + (file.folderId ? file.folderId : ""));
-};
+})
 
-exports.downloadFile = async (req, res) => {
+exports.downloadFile = asyncHandler(async (req, res) => {
   const userId = req.user.id;
   const fileId = req.params.fileId;
 
   const file = await fileQuery.getFile(userId, fileId);
+
+  if (!file) {
+    const error = new Error("File not found.")
+    error.status = 404;
+    throw error;
+  }
 
   const { data, error } = await supabase.storage
     .from("uploads")
@@ -223,4 +274,4 @@ exports.downloadFile = async (req, res) => {
   res.setHeader("Content-Length", buffer.length);
 
   res.send(buffer);
-};
+});
